Rename search attribute filter component
