+++ src/lib/components/standalone-activities/activities-summary-filter-bar/filter-bar.svelte
@@ -1,7 +1,7 @@
 <svelte:options runes />
 
 <script lang="ts">
-  import FilterBar from '$lib/components/shared-search-attribute-filter/filter-bar.svelte';
+  import FilterBar from '$lib/components/search-attribute-filter/filter-bar.svelte';
   import { activitiesQuery, activityRefresh } from '$lib/stores/activities';
   import { activityFilters } from '$lib/stores/filters';
   import { activitySearchAttributeOptions } from '$lib/stores/search-attributes';
